+++ cmd/resources.go
@@ -1,0 +1,34 @@
+// Copyright © 2017 G. Hussain Chinoy <ghchinoy@gmail.com>
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+)
+
+// resourcesCmd represents the resources command
+var resourcesCmd = &cobra.Command{
+	Use:   "resources",
+	Short: "Manage common resources",
+	Long:  `List, add, remove and inspect common resource objects.`,
+}
+
+func init() {
+	RootCmd.AddCommand(resourcesCmd)
+
+	resourcesCmd.PersistentFlags().StringVar(&profile, "profile", "default", "profile name")
+	resourcesCmd.PersistentFlags().BoolVarP(&outputJSON, "json", "j", false, "output as json")
+	resourcesCmd.PersistentFlags().BoolVarP(&showCurl, "curl", "c", false, "show curl command")
+}